+++ pyxelrest/pyxelrestgenerator.py
@@ -26,7 +26,7 @@ else:
     from imp import reload
 
 
-def user_defined_functions(loaded_services, flattenize=True):
+def user_defined_functions(loaded_services, pyxelrest_config, flattenize=True):
     """
     Create xlwings User Defined Functions according to user_defined_functions template.
     :return: A string containing python code with all xlwings UDFs.
@@ -39,6 +39,7 @@ def user_defined_functions(loaded_services, flattenize=True):
     return renderer.get_template('user_defined_functions.jinja2').render(
         current_utc_time=datetime.datetime.utcnow().isoformat(),
         services=loaded_services,
+        pyxelrest_config=pyxelrest_config,
         modified_parameters={value: key for key, value in vba.vba_restricted_keywords.items()},
         support_pandas=swagger.support_pandas(),
         support_ujson=support_ujson(),
@@ -61,11 +62,11 @@ def generate_user_defined_functions(output='user_defined_functions.py', flatteni
     :param flattenize: Set to False if you want the JSON dictionary as result of your UDF call.
     :return: None
     """
-    services = swagger.load_services()
+    services, pyxelrest_config = swagger.load_services()
     logging.debug('Generating user defined functions.')
     with open(os.path.join(os.path.dirname(__file__), output), 'w', encoding='utf-8') \
             as generated_file:
-        generated_file.write(user_defined_functions(services, flattenize))
+        generated_file.write(user_defined_functions(services, pyxelrest_config, flattenize))
 
 
 def reload_user_defined_functions():